rename dask to scheduler, add future imports
