+++ distributed3/client.py
@@ -1,3 +1,5 @@
+from __future__ import print_function, division, absolute_import
+
 import random
 import uuid
 from itertools import count, cycle
