+++ distributed3/dask.py
@@ -1,4 +1,4 @@
-from __future__ import division, print_function
+from __future__ import print_function, division, absolute_import
 
 from math import ceil
 from time import time
