+++ distributed3/utils.py
@@ -1,3 +1,4 @@
+from __future__ import print_function, division, absolute_import
 
 import socket
