+++ distributed3/worker.py
@@ -1,4 +1,4 @@
-from __future__ import print_function, division
+from __future__ import print_function, division, absolute_import
 
 from concurrent.futures import ThreadPoolExecutor
 from multiprocessing.pool import ThreadPool
